// src/modules/auth/domain/enums/auth-exception-code.ts
export enum AuthExceptionCode {
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  EMAIL_ALREADY_TAKEN = 'EMAIL_ALREADY_TAKEN',
  USERNAME_ALREADY_TAKEN = 'USERNAME_ALREADY_TAKEN',
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  USER_DELETED = 'USER_DELETED',
  USER_ALREADY_DELETED = 'USER_ALREADY_DELETED',
  USER_MUST_BE_ACTIVE = 'USER_MUST_BE_ACTIVE',
  USER_MUST_BE_DISABLED = 'USER_MUST_BE_DISABLED',
  INVALID_USER_STATUS = 'INVALID_USER_STATUS',
  USER_GROUP_NOT_FOUND = 'USER_GROUP_NOT_FOUND',
  ROLE_NOT_FOUND = 'ROLE_NOT_FOUND',
  USER_ALREADY_IN_GROUP = 'USER_ALREADY_IN_GROUP',
  USER_NOT_IN_GROUP = 'USER_NOT_IN_GROUP',
  ROLE_ALREADY_IN_GROUP = 'ROLE_ALREADY_IN_GROUP',
  ROLE_NOT_IN_GROUP = 'ROLE_NOT_IN_GROUP',
  USER_GROUP_NAME_ALREADY_TAKEN = 'USER_GROUP_NAME_ALREADY_TAKEN',
  EXTERNAL_AUTHENTICATION_ERROR = 'EXTERNAL_AUTHENTICATION_ERROR',
}