// src/modules/auth/domain/enums/user-event-type.ts
/**
 * User domain event types
 */
export enum UserEventType {
  REGISTERED = 'USER_REGISTERED',
  UPDATED = 'USER_UPDATED',
  STATUS_CHANGED = 'USER_STATUS_CHANGED',
  DELETED = 'USER_DELETED',
  ADDED_TO_USER_GROUP = 'USER_ADDED_TO_USER_GROUP',
  REMOVED_FROM_USER_GROUP = 'USER_REMOVED_FROM_USER_GROUP',
}